+++ test/api/test_libraries.py
@@ -1,6 +1,5 @@
 from base import api
-
-from .helpers import (
+from base.populators import (
     LibraryPopulator,
     TestsDatasets,
     wait_on_state
